Resolvers missing from World
